requirements.txt
requests~=2.32.4
django-environ~=0.12.0
Flask~=3.1.1
python-dotenv~=1.1.0